consolidate rejection test into s3-list-objects.test.ts; skip in-process tests when HTTP_PROXY is inherited
